video-summary-mcp/requirements.txt
google-genai>=1.5.0
fastapi>=0.80.0
uvicorn[standard]>=0.18.0
pydantic>=2.0.0,<3.0.0
fastapi-mcp